Split opendistro-tar-install.sh into two pieces: opendistro-onetime-setup.sh and opendistro-start.sh
This change is to ensure that one-time activities, such as installing the demo securityo config, happen once. This is especially useful for Docker images where we can run the one-time script as part of the Docker build.

opendistro-tar-install.sh invokes opendistro-onetime-setup followed by opendistro-start so that existing scripts that use opendistro-tar-install will continue to work the same as they did before this change.